requirements.txt
PyYAML>=6.0
jsonschema>=4.0.0
pytest>=7.0.0